// prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AdminLevel = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  MODERATOR: 'MODERATOR',
  USER: 'USER'
} as const

export type AdminLevel = (typeof AdminLevel)[keyof typeof AdminLevel]


export const Roles = {
  user: 'user',
  admin: 'admin'
} as const

export type Roles = (typeof Roles)[keyof typeof Roles]


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  cancelled: 'cancelled',
  shipped: 'shipped',
  delivered: 'delivered',
  refunded: 'refunded'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const CartStatus = {
  active: 'active',
  abandoned: 'abandoned',
  checked_out: 'checked_out',
  expired: 'expired'
} as const

export type CartStatus = (typeof CartStatus)[keyof typeof CartStatus]